CommonLib, ClientLib, Data: 부동 소수점 오차를 고려해 체력, 대미지, 사거리, 이동 속도를 정수 기반으로 변경
기존에는 float를 사용했는데, byte, ushort, uint를 사용해 정수 기반으로 변경하였음.

이동 방식도 단계적으로 정수 기반으로 변환할 예정.

+++ Assets/Script/ClientLib/ClientBattleMapSimulator.cs
@@ -140,7 +140,7 @@ namespace Script.ClientLib
             attacker.OnStartAttack();
         }
 
-        public void OnEntityGetDamage(uint entityId, float damage)
+        public void OnEntityGetDamage(uint entityId, uint damage)
         {
             _entityViews.TryGetValue(entityId, out var entityView);
